rename vorfaelleundvorkommnisseStore.ts to ereignisStore.ts
