+++ wls-gui-wahllokalsystem/src/stores/ereignisStore.ts
